Upload Data and CloudPareto Case Study
CloudPareto’s product portfolio supports two kinds of users: buyers and sellers. Our first product (Product B) provides a way for buyers to purchase items while our second product (Product S) allows sellers to identify new ways to pitch to the buyers. Product S is a social media type platform that allows sellers to:

- Post and interact with posts from other sellers as well as buyers
- Message other buyers and sellers
- Partner with other sellers to create stronger sales pitches
- Search for buyer requirements
- Connect directly to Product B to submit ideas directly to buyers

One of the ways the Analytics Team will support Product S is through building recommendation systems to provide users with more actionable insights.

For this case, you have been provided a list of seller posts that can be found in the Product S product feed. This information includes the Seller ID, the date of the post, the seller’s state, and the post text. We would like you to build a model to help sellers identify other sellers that they may want to partner with based on the post information.

During the case interview, you will be asked to discuss/show:
- Approach and why you chose it
- Insights gathered from the exercise
- What additional data elements you would consider to improve model accuracy
- Code (it will not be evaluated on cleanliness or production readiness)

We recognize this is a limited dataset. Our aim is not to have you develop a model with high accuracy in this short period but rather see your approach. Please spend no more than 3 hours on this.